requirements.txt
PyYAML
mock
>>>>>>> feature/alanis
feedparser
astroid==2.3.3
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.1
pluggy==0.13.1
py==1.8.1
pylint==2.4.4
pyparsing==2.4.6
pytest==5.3.5
requests==2.23.0
six==1.14.0
urllib3==1.25.8
wcwidth==0.1.8
wrapt==1.11.2
